renaming, mod, ignore same-case concurrence from aspell suggestions
